+++ example/powerStationPlain.py
@@ -1,0 +1,9 @@
+import cadquery as cq
+from cqspoolterrain import PowerStation, SpoolCladdingGreebled
+
+bp_power = PowerStation()
+
+bp_power.make()
+power = bp_power.build()
+#show_object(power)
+cq.exporters.export(power,f"stl/powerStation.stl")