+++ changes.md
@@ -1,9 +1,15 @@
 # CQSpoolTerrain changelog
 
 ## Main wip
+## 0.0.4
+* Made SpoolCladdingGreebled class
+* Made SpoolCladdingGreebledUnique class
+* Added examples
+* Added stls
+
 ## 0.0.3
 * Made SpoolCladding class.
-* Intrgrated SpoolCladding into PowerStation class
+* Integrated SpoolCladding into PowerStation class
 
 ## 0.0.2
 * Upped cqterrain verison 0.1.8